refactor(voice): Refactor voice service implementations and enhance UI volume visualization
- Specialized service implementations (Local Whisper, Piper, OpenAI) are introduced for decoupled voice functionality.
- Refactor TTS and STT services into dedicated, specialized classes.
- Apply perceptual volume boosting to voice waveform samples in the UI for better visual fidelity.
