+++ desktop-shared/src/main/kotlin/io/askimo/ui/voice/impl/OpenAiTextToSpeechService.kt
@@ -8,6 +8,7 @@ import dev.langchain4j.model.openai.OpenAiTextToSpeechModel
 import io.askimo.core.config.VoiceConfig
 import io.askimo.core.config.VoiceProvider
 import io.askimo.core.logging.logger
+import io.askimo.core.util.createJdkHttpClientBuilder
 import io.askimo.ui.voice.TextToSpeechFactory
 import io.askimo.ui.voice.TextToSpeechService
 import io.askimo.ui.voice.VoiceAudioFormat
@@ -36,6 +37,7 @@ class OpenAiTextToSpeechService(private val config: VoiceConfig) : TextToSpeechS
 
         try {
             val tts = OpenAiTextToSpeechModel.builder()
+                .httpClientBuilder(createJdkHttpClientBuilder())
                 .apiKey(apiKey)
                 .modelName(config.ttsModel.ifBlank { "tts-1" })
                 .voice(config.ttsVoice.ifBlank { "alloy" })
@@ -51,49 +53,4 @@ class OpenAiTextToSpeechService(private val config: VoiceConfig) : TextToSpeechS
 object OpenAiTextToSpeechFactory : TextToSpeechFactory {
     override val provider: VoiceProvider = VoiceProvider.OPENAI
     override fun create(config: VoiceConfig): TextToSpeechService = OpenAiTextToSpeechService(config)
-}
-
-/**
- * Text-to-speech via a user-hosted Piper HTTP server exposing an OpenAI-compatible
- * `/v1/audio/speech` endpoint (e.g. `piper-http` OpenAI-compatible wrappers). Free, no API key
- * required, runs fully offline.
- *
- * Uses langchain4j's [OpenAiTextToSpeechModel] pointed at a custom [VoiceConfig.localTtsEndpoint]
- * base URL (default `http://localhost:5000`) — the same "OpenAI-compatible" pattern used by
- * [io.askimo.core.providers.openaicompatible.OpenAiCompatibleModelFactory] for chat/embedding
- * models and by [LocalWhisperSpeechToTextService] for local speech-to-text.
- */
-class PiperTextToSpeechService(private val config: VoiceConfig) : TextToSpeechService {
-    private val log = logger<PiperTextToSpeechService>()
-
-    override val outputFormat: VoiceAudioFormat = VoiceAudioFormat.WAV
-
-    override suspend fun synthesize(text: String): ByteArray = withContext(Dispatchers.IO) {
-        val baseUrl = config.localTtsEndpoint.trimEnd('/')
-        if (baseUrl.isBlank()) {
-            throw VoiceServiceException("Local Piper endpoint is not configured. Set it in Settings > Voice.")
-        }
-
-        try {
-            val tts = OpenAiTextToSpeechModel.builder()
-                .baseUrl(baseUrl)
-                .apiKey(config.openAiApiKey.ifBlank { "not-needed" })
-                .modelName(config.ttsModel.ifBlank { "tts-1" })
-                .voice(config.ttsVoice.ifBlank { "alloy" })
-                .build()
-
-            tts.synthesize(text).audio().binaryData()
-        } catch (e: Exception) {
-            log.warn("Local Piper TTS request failed", e)
-            throw VoiceServiceException(
-                "Could not reach local Piper server at $baseUrl. Is it running? (${e.message})",
-                e,
-            )
-        }
-    }
-}
-
-object PiperTextToSpeechFactory : TextToSpeechFactory {
-    override val provider: VoiceProvider = VoiceProvider.LOCAL_PIPER
-    override fun create(config: VoiceConfig): TextToSpeechService = PiperTextToSpeechService(config)
 }
